State Need For Manual "build directory" In ReadMe
The present ReadMe Getting Started instructions do not indicate that a new build directory needs to be created by the user.  autoget.sh should be run in the root of the project, whereas the configure command needs to be run in a sub-directory.  Expand the instructions to show creating and navigating to a build directory and then markup commands.