stop creating temp files that aren't cleaned up
When I got 770f14e246244d867a78aca3847dc4bc19d9483c merged I stored the
return code from the test binary in a temp file and then deleted the
temp file.
But that temp file was created every time MakefileWorker.mk was sourced,
not just when the targets were run.

This commit changes it so the temp file is only created if the relevant
targets are being run so it will always get deleted.